Removed Util and replaces some code with generic
