+++ mail/src/main/java/org/jboss/as/mail/extension/MailServerWriteAttributeHandler.java
@@ -8,12 +8,17 @@ import org.jboss.as.controller.PathAddress;
 import org.jboss.as.controller.registry.ManagementResourceRegistration;
 import org.jboss.dmr.ModelNode;
 
+import static org.jboss.as.mail.extension.MailServerDefinition.OUTBOUND_SOCKET_BINDING_REF;
+import static org.jboss.as.mail.extension.MailServerDefinition.PASSWORD;
+import static org.jboss.as.mail.extension.MailServerDefinition.SSL;
+import static org.jboss.as.mail.extension.MailServerDefinition.USERNAME;
+
 /**
  * @author Tomaz Cerar
  * @created 22.12.11 18:31
  */
 public class MailServerWriteAttributeHandler extends AbstractWriteAttributeHandler<Boolean> {
-    private static final AttributeDefinition[] ATTRIBUTES = {MailServerDefinition.OUTBOUND_SOCKET_BINDING_REF, MailServerDefinition.SSL, MailServerDefinition.USERNAME, MailServerDefinition.PASSWORD};
+    private static final AttributeDefinition[] ATTRIBUTES = {OUTBOUND_SOCKET_BINDING_REF, SSL, USERNAME, PASSWORD};
     static final MailServerWriteAttributeHandler INSTANCE = new MailServerWriteAttributeHandler();
 
     private MailServerWriteAttributeHandler() {
@@ -28,8 +33,8 @@ public class MailServerWriteAttributeHandler extends AbstractWriteAttributeHandl
 
     @Override
     protected boolean applyUpdateToRuntime(OperationContext context, ModelNode operation, String attributeName, ModelNode resolvedValue, ModelNode currentValue, HandbackHolder<Boolean> handbackHolder) throws OperationFailedException {
-        final ModelNode model = context.readResource(PathAddress.EMPTY_ADDRESS).getModel();
-        boolean handback = applyModelToRuntime(context, operation, attributeName, model);
+        //final ModelNode model = context.readResource(PathAddress.EMPTY_ADDRESS).getModel();
+        boolean handback = true;
         handbackHolder.setHandback(handback);
         return handback;
     }
@@ -43,10 +48,5 @@ public class MailServerWriteAttributeHandler extends AbstractWriteAttributeHandl
         } // else we didn't update the runtime in applyUpdateToRuntime
     }
 
-    private boolean applyModelToRuntime(OperationContext context, ModelNode operation, String attributeName, ModelNode model) throws OperationFailedException {
 
-        boolean reloadRequired = true;
-
-        return reloadRequired;
-    }
 }
