+++ mail/src/main/java/org/jboss/as/mail/extension/MailSessionAdd.java
@@ -23,6 +23,8 @@ import org.jboss.msc.value.ImmediateValue;
 
 import java.util.List;
 
+import static org.jboss.as.mail.extension.MailSubsystemModel.*;
+
 /**
  * @author Tomaz Cerar
  * @created 27.7.11 0:55
@@ -39,8 +41,10 @@ public class MailSessionAdd extends AbstractAddStepHandler {
      * {@inheritDoc}
      */
     @Override
-    protected void populateModel(ModelNode existingModel, ModelNode newModel) throws OperationFailedException {
-        Util.copyModel(existingModel,newModel,MailSubsystemModel.JNDI_NAME,MailSubsystemModel.DEBUG,MailSubsystemModel.FROM);
+    protected void populateModel(ModelNode operation, ModelNode model) throws OperationFailedException {
+        MailSessionDefinition.JNDI_NAME.validateAndSet(operation, model);
+        MailSessionDefinition.DEBUG.validateAndSet(operation, model);
+        MailSessionDefinition.FROM.validateAndSet(operation, model);
     }
 
 
@@ -66,11 +70,11 @@ public class MailSessionAdd extends AbstractAddStepHandler {
      */
     @Override
     protected void performRuntime(OperationContext context, ModelNode operation, ModelNode model, ServiceVerificationHandler verificationHandler, List<ServiceController<?>> controllers) throws OperationFailedException {
-        final String jndiName = Util.getJndiName(operation);
+        final String jndiName = getJndiName(operation);
         final ServiceTarget serviceTarget = context.getServiceTarget();
 
         ModelNode fullTree = Resource.Tools.readModel(context.readResource(PathAddress.EMPTY_ADDRESS));
-        final MailSessionConfig config = Util.from(context, fullTree);
+        final MailSessionConfig config = from(context, fullTree);
         final MailSessionService service = new MailSessionService(config);
         final ServiceName serviceName = SERVICE_NAME_BASE.append(jndiName);
         final ServiceBuilder<?> mailSessionBuilder = serviceTarget.addService(serviceName, service);
@@ -115,9 +119,25 @@ public class MailSessionAdd extends AbstractAddStepHandler {
                 .addListener(verificationHandler);
         controllers.add(mailSessionBuilder.install());
         controllers.add(binderBuilder.install());
-
-
     }
+    /**
+         * Extracts the raw JNDI_NAME value from the given model node, and depending on the value and
+         * the value of any USE_JAVA_CONTEXT child node, converts the raw name into a compliant jndi name.
+         *
+         * @param modelNode the model node; either an operation or the model behind a mail session resource
+         * @return the compliant jndi name
+         */
+        public static String getJndiName(final ModelNode modelNode) {
+            final String rawJndiName = modelNode.require(JNDI_NAME).asString();
+            final String jndiName;
+            if (!rawJndiName.startsWith("java:")) {
+                jndiName = "java:jboss/mail/" + rawJndiName;
+            } else {
+                jndiName = rawJndiName;
+            }
+            return jndiName;
+        }
+
 
     private void addOutboundSocketDependency(MailSessionService service, ServiceBuilder<?> mailSessionBuilder, MailSessionServer server) {
         if (server != null) {
@@ -125,5 +145,42 @@ public class MailSessionAdd extends AbstractAddStepHandler {
             mailSessionBuilder.addDependency(OutboundSocketBinding.OUTBOUND_SOCKET_BINDING_BASE_SERVICE_NAME.append(ref),
                     OutboundSocketBinding.class, service.getSocketBindingInjector(ref));
         }
+    }
+
+    static MailSessionConfig from(final OperationContext operationContext, final ModelNode model) throws OperationFailedException {
+        MailSessionConfig cfg = new MailSessionConfig();
+        cfg.setJndiName(model.require(JNDI_NAME).asString());
+        cfg.setDebug(model.get(DEBUG).asBoolean(false));
+        cfg.setFrom(model.get(FROM).asString());
+
+        if (model.hasDefined(SERVER_TYPE)) {
+            ModelNode server = model.get(SERVER_TYPE);
+            if (server.hasDefined(SMTP)) {
+                cfg.setSmtpServer(readServerConfig(operationContext, server.get(SMTP)));
+            }
+            if (server.hasDefined(POP3)) {
+                cfg.setPop3Server(readServerConfig(operationContext, server.get(POP3)));
+            }
+            if (server.hasDefined(IMAP)) {
+                cfg.setImapServer(readServerConfig(operationContext, server.get(IMAP)));
+            }
+        }
+        return cfg;
+    }
+
+    private static MailSessionServer readServerConfig(final OperationContext operationContext, final ModelNode model) throws OperationFailedException {
+        final String socket = model.require(OUTBOUND_SOCKET_BINDING_REF).asString();
+        final Credentials credentials = readCredentials(operationContext, model);
+        boolean ssl = model.get(SSL).asBoolean(false);
+        return new MailSessionServer(socket, credentials, ssl);
+    }
+
+    private static Credentials readCredentials(final OperationContext operationContext, final ModelNode model) throws OperationFailedException {
+        if (model.has(USER_NAME)) {
+            String un = model.get(USER_NAME).asString();
+            String pw = operationContext.resolveExpressions((model.get(PASSWORD))).asString();
+            return new Credentials(un, pw);
+        }
+        return null;
     }
 }
