+++ mail/src/main/java/org/jboss/as/mail/extension/MailSessionDefinition.java
@@ -24,22 +24,23 @@ public class MailSessionDefinition extends SimpleResourceDefinition {
     protected static final SimpleAttributeDefinition JNDI_NAME =
             new SimpleAttributeDefinitionBuilder(MailSubsystemModel.JNDI_NAME, ModelType.STRING, true)
                     .setAllowExpression(true)
-                    .setFlags(AttributeAccess.Flag.RESTART_RESOURCE_SERVICES)
+                    .setRestartAllServices()
                     .build();
     protected static final SimpleAttributeDefinition FROM =
             new SimpleAttributeDefinitionBuilder(MailSubsystemModel.FROM, ModelType.STRING, true)
                     .setAllowExpression(true)
-                    .setFlags(AttributeAccess.Flag.RESTART_RESOURCE_SERVICES)
+                    .setRestartAllServices()
                     .build();
     protected static final SimpleAttributeDefinition DEBUG =
             new SimpleAttributeDefinitionBuilder(MailSubsystemModel.DEBUG, ModelType.BOOLEAN, true)
                     .setAllowExpression(true)
-                    .setFlags(AttributeAccess.Flag.RESTART_RESOURCE_SERVICES)
+                    .setRestartAllServices()
                     .build();
 
     @Override
     public void registerAttributes(final ManagementResourceRegistration rootResourceRegistration) {
         MailSessionWriteAttributeHandler.INSTANCE.registerAttributes(rootResourceRegistration);
     }
+
 
 }
